bug: Fix STDIN/STDOUT buffer issue on code generation with code formatters
There was a problem that it cannot generate code If the input size is too big, because of the buffer of the OS.
This commit fixes that problem by using `exec.Cmd#StdinPipe()` and removes some redundant/unsophisticated code.

As far as I checked, this problem appeared on only Linux.

Signed-off-by: moznion <moznion@mail.moznion.net>